Move customize_compiler back to distutils.sysconfig as it'll be removedtogether with the module.
